+++ internal/models/patients.go
@@ -153,16 +153,42 @@ func (m *PatientModel) GetAllByMedication(medication string) ([]*Patient, error)
 	return patients, nil
 }
 
-func (m *PatientModel) Update(id int, ucn string, firstName string, lastName string, phone string, height int, weight int, medication string, note string, approved bool, firstCont bool) error {
-	stmt := "UPDATE patients SET ucn = ?, first_name = ?, last_name = ?, phone_number = ?, height = ?, weight = ?, medication = ?, note = ?, approved = ?, first_continuation = ? WHERE id = ?"
+func (m *PatientModel) Update(id int, ucn string, firstName string, lastName string, phone string, height int, weight int, medication string, note string, approved bool, firstCont bool, userId int) error {
+	stmt := "UPDATE patients SET ucn = ?, first_name = ?, last_name = ?, phone_number = ?, height = ?, weight = ?, medication = ?, note = ?, approved = ?, first_continuation = ? WHERE id = ? && user_id = ?"
 
-	_, err := m.DB.Exec(stmt, ucn, firstName, lastName, phone, height, weight, medication, note, approved, firstCont, id)
-	return err
+	res, err := m.DB.Exec(stmt, ucn, firstName, lastName, phone, height, weight, medication, note, approved, firstCont, id, userId)
+	if err != nil {
+		return err
+	}
+
+	rows, err := res.RowsAffected()
+	if err != nil {
+		return err
+	}
+
+	if rows == 0 {
+		return ErrUnauthorizedAction
+	}
+
+	return nil
 }
 
-func (m *PatientModel) Delete(id int) error {
-	stmt := "DELETE FROM patients WHERE id = ?"
+func (m *PatientModel) Delete(id int, userId int) error {
+	stmt := "DELETE FROM patients WHERE id = ? && user_id = ?"
 
-	_, err := m.DB.Exec(stmt, id)
-	return err
+	res, err := m.DB.Exec(stmt, id, userId)
+	if err != nil {
+		return err
+	}
+
+	rows, err := res.RowsAffected()
+	if err != nil {
+		return err
+	}
+
+	if rows == 0 {
+		return ErrUnauthorizedAction
+	}
+
+	return nil
 }
